+++ dds/src/dds_async/domain_participant.rs
@@ -1,11 +1,11 @@
-use std::sync::Arc;
-
+use super::{
+    condition::StatusConditionAsync, domain_participant_listener::DomainParticipantListenerAsync,
+    publisher::PublisherAsync, publisher_listener::PublisherListenerAsync,
+    subscriber::SubscriberAsync, subscriber_listener::SubscriberListenerAsync, topic::TopicAsync,
+    topic_listener::TopicListenerAsync,
+};
 use crate::{
     builtin_topics::{ParticipantBuiltinTopicData, TopicBuiltinTopicData},
-    data_representation_builtin_endpoints::{
-        discovered_topic_data::DCPS_TOPIC,
-        spdp_discovered_participant_data::{SpdpDiscoveredParticipantData, DCPS_PARTICIPANT},
-    },
     domain::domain_participant_factory::DomainId,
     implementation::{
         actor::{Actor, ActorAddress},
@@ -19,6 +19,10 @@ use crate::{
             subscriber_actor::{self, SubscriberActor},
             topic_actor::{self, TopicActor},
         },
+        data_representation_builtin_endpoints::{
+            discovered_topic_data::DCPS_TOPIC,
+            spdp_discovered_participant_data::{SpdpDiscoveredParticipantData, DCPS_PARTICIPANT},
+        },
         runtime::{executor::ExecutorHandle, timer::TimerHandle},
     },
     infrastructure::{
@@ -30,13 +34,7 @@ use crate::{
     },
     topic_definition::type_support::{DdsHasKey, DdsKey, DdsTypeXml, DynamicTypeInterface},
 };
-
-use super::{
-    condition::StatusConditionAsync, domain_participant_listener::DomainParticipantListenerAsync,
-    publisher::PublisherAsync, publisher_listener::PublisherListenerAsync,
-    subscriber::SubscriberAsync, subscriber_listener::SubscriberListenerAsync, topic::TopicAsync,
-    topic_listener::TopicListenerAsync,
-};
+use std::sync::Arc;
 
 /// Async version of [`DomainParticipant`](crate::domain::domain_participant::DomainParticipant).
 #[derive(Clone)]
