+++ dds/src/dds_async/subscriber.rs
@@ -1,5 +1,9 @@
+use super::{
+    condition::StatusConditionAsync, data_reader::DataReaderAsync,
+    data_reader_listener::DataReaderListenerAsync, domain_participant::DomainParticipantAsync,
+    subscriber_listener::SubscriberListenerAsync, topic::TopicAsync,
+};
 use crate::{
-    data_representation_builtin_endpoints::discovered_reader_data::DCPS_SUBSCRIPTION,
     implementation::{
         actor::{Actor, ActorAddress},
         actors::{
@@ -10,6 +14,7 @@ use crate::{
             subscriber_actor::{self, SubscriberActor},
             topic_actor::{self, TopicActor},
         },
+        data_representation_builtin_endpoints::discovered_reader_data::DCPS_SUBSCRIPTION,
     },
     infrastructure::{
         error::{DdsError, DdsResult},
@@ -17,12 +22,6 @@ use crate::{
         qos::{DataReaderQos, QosKind, SubscriberQos, TopicQos},
         status::{SampleLostStatus, StatusKind},
     },
-};
-
-use super::{
-    condition::StatusConditionAsync, data_reader::DataReaderAsync,
-    data_reader_listener::DataReaderListenerAsync, domain_participant::DomainParticipantAsync,
-    subscriber_listener::SubscriberListenerAsync, topic::TopicAsync,
 };
 
 /// Async version of [`Subscriber`](crate::subscription::subscriber::Subscriber).
