+++ dds/src/implementation/data_representation_builtin_endpoints/discovered_writer_data.rs
@@ -157,6 +157,43 @@ impl DdsSerialize for DiscoveredWriterData {
     }
 }
 
+impl<'de> DdsDeserialize<'de> for PublicationBuiltinTopicData {
+    fn deserialize_data(serialized_data: &'de [u8]) -> DdsResult<Self> {
+        let pl_deserializer = ParameterListCdrDeserializer::new(serialized_data)?;
+        Ok(Self {
+            key: pl_deserializer.read(PID_ENDPOINT_GUID)?,
+            // Default value is a deviation from the standard and is used for interoperability reasons:
+            participant_key: pl_deserializer
+                .read_with_default(PID_PARTICIPANT_GUID, Default::default())?,
+            topic_name: pl_deserializer.read(PID_TOPIC_NAME)?,
+            type_name: pl_deserializer.read(PID_TYPE_NAME)?,
+            durability: pl_deserializer.read_with_default(PID_DURABILITY, Default::default())?,
+            deadline: pl_deserializer.read_with_default(PID_DEADLINE, Default::default())?,
+            latency_budget: pl_deserializer
+                .read_with_default(PID_LATENCY_BUDGET, Default::default())?,
+            liveliness: pl_deserializer.read_with_default(PID_LIVELINESS, Default::default())?,
+            reliability: pl_deserializer
+                .read_with_default(PID_RELIABILITY, DEFAULT_RELIABILITY_QOS_POLICY_DATA_WRITER)?,
+            lifespan: pl_deserializer.read_with_default(PID_LIFESPAN, Default::default())?,
+            user_data: pl_deserializer.read_with_default(PID_USER_DATA, Default::default())?,
+            ownership: pl_deserializer.read_with_default(PID_OWNERSHIP, Default::default())?,
+            ownership_strength: pl_deserializer
+                .read_with_default(PID_OWNERSHIP_STRENGTH, Default::default())?,
+            destination_order: pl_deserializer
+                .read_with_default(PID_DESTINATION_ORDER, Default::default())?,
+            presentation: pl_deserializer
+                .read_with_default(PID_PRESENTATION, Default::default())?,
+            partition: pl_deserializer.read_with_default(PID_PARTITION, Default::default())?,
+            topic_data: pl_deserializer.read_with_default(PID_TOPIC_DATA, Default::default())?,
+            group_data: pl_deserializer.read_with_default(PID_GROUP_DATA, Default::default())?,
+            xml_type: pl_deserializer
+                .read_with_default(PID_TYPE_REPRESENTATION, Default::default())?,
+            representation: pl_deserializer
+                .read_with_default(PID_DATA_REPRESENTATION, Default::default())?,
+        })
+    }
+}
+
 impl<'de> DdsDeserialize<'de> for DiscoveredWriterData {
     fn deserialize_data(serialized_data: &'de [u8]) -> DdsResult<Self> {
         let pl_deserializer = ParameterListCdrDeserializer::new(serialized_data)?;
