+++ dds/src/implementation/mod.rs
@@ -1,4 +1,5 @@
 pub mod actor;
 pub mod actors;
+pub mod data_representation_builtin_endpoints;
 pub mod data_representation_inline_qos;
 pub mod runtime;
